MUST NOT accept subsequent OTPs, after valid OTP verification
This is in compliance with RFC 6238, Section 5.2, final paragraph.

The TOTP class has been amended to have a hash of consumed or
"burned" one time pads at their associated time step. For a given
OTP, it is added to the hash if successfully validated and isn't
already in the hash for that timestep.

Refs #44.